handle case where consistency with callback is used
The same flow as with the other consistency levels is adopted for now.
In this commit, wait and get operations are split. In between, C*
should return the first piece of data in hand to the client. After
that, the usual mismatch check is invoked. If a mismatch occurs, data
resolver is invoked and the result is passed in a second response to
the client, if not, a lightweight ack is still needed to inform the
client to go ahead.